requirements.txt
Django==5.2.1
django-isbn-field==0.5.3
djangorestframework==3.16.0
django-cors-headers==4.7.0